Fix "Uncaught TypeError" in the JS console.
The status toggle controller JS file was using some of the Hotwire concepts (turbo:load) but was not defining a 'Controller', in the Stimulus sense. The way it was implemented caused it to try and load on every page, even when no status dropdown existed. I converted the code into a Stimulus controller and attached it to the appropriate page. If this logic is only used on a single form it might be easier to understand in the page itself instead of as a Controller.