Fix "Uncaught TypeError" in the JS console. (#48)
* Fix "Uncaught TypeError" in the JS console.

The status toggle controller JS file was using some of the Hotwire concepts (turbo:load) but was not defining a 'Controller', in the Stimulus sense. The way it was implemented caused it to try and load on every page, even when no status dropdown existed. I converted the code into a Stimulus controller and attached it to the appropriate page. If this logic is only used on a single form it might be easier to understand in the page itself instead of as a Controller.

* Add name and date to DEVELOPERS.md

* Remove to two non-Controller JS files ending in _controller. The logic in each was only on a single page so it was simpler to inline it. I tested it with fresh loads and navigation to verify it works in turbo naviagtion correctly. I also changed the address form to use a consisternt usage and cleaned up some browser console warnings.

* Cleanup the list of checkboxes controller by the javascript. Getting all of the rows as siblings cleaned up some odd spacing.

* Update stimulus manifest to reflect recent changes

---------

Co-authored-by: Bob Gatewood <bgatewood@users.noreply.github.com>