Remove project_id requirements for fisher autocomplete. (#56)
This change moves the search action from /manage/projects/:project_id/finishers/search(.:format) to /manage/projects/finishers/search(.:format) and the :project_id was not used in the search action.

The finisher autocomplete route was set to require a :project_id. When creating a new project the path method was being called with a valid Project object, but one which had no id (from Project.new). This caused the path method to raise an error because it was trying to use the id of the Project object as the :project_id parameter. Since it was not used it seemed easiest to remove the requirement for the :project_id parameter.