Update assignment form UX
Correct the turbo streams UX issue. Before this updates to inline forms
were invisible. It made it feel like the work was not completed. Now the
form shows "Saved" upon save. The three column layout on the project
manage page was making it difficult to show the new information so notes
and assignments were moved to the bottom.