Devcontainer Setup issues
I ran into a few issues setting up the VSCode devcontainer so I
went ahead and fixed them here. The yarnrc.yml fixes an issue where
the Linux container on a Mac complained about the binary format. The
Procfile was needed for bin/dev to be accessible outside of the
container, in the browser.