Create Assignment Statuses (#100)
* Create assignment status data model

Add the status field to the Assignment model with validation and tests.
This also adds a test for one existing helper scope-like method.

* Fix bug when Finisher missing joined_on

Signed-off-by: Matt Sanford <matt@mzsanford.com>

* Assignment status manage UI

Add the assignment status to the project display. Also allow status
change and delete from the project edit view. There was a broken link to
the assignment manage index so I added that page as well.

* Update assignment form UX

Correct the turbo streams UX issue. Before this updates to inline forms
were invisible. It made it feel like the work was not completed. Now the
form shows "Saved" upon save. The three column layout on the project
manage page was making it difficult to show the new information so notes
and assignments were moved to the bottom.

* Allow assignment status update on project show page

* Remove form actions infavor of save on change

---------

Signed-off-by: Matt Sanford <matt@mzsanford.com>
Co-authored-by: Robert R. Meyer <20780405+robert-r-meyer@users.noreply.github.com>