Assignment status validation follow ups (#133)
* Assignment status blank handling

Allow an empty string to be passed in as the Assignment#status. This is
treated the same as nil. This is needed to accommodate any Assignment
that existed prior to the introduction of the status field.

* Assignment status normalization

Keep the database cleaner by coercing the status to nil if a blank
value is passed in. This way we don't need to handle the nil/blank at
display time.

* Assignment status "saved" label consistency

Change the "saved" label on the assignment form from using updated_at
with an arbitrary 5 second window to using saved_change_to_attribute?
from ActiveModel::Dirty. This prevents showing "saved" if the page is
refreshed within 5 seconds of the last save or right after create.

* Add controller tests for Manage::AssignmentsController#update