First pass of search as a Concern
Create a new ActiveModel::Concern top handle the parts of search that
apply to multiple models. This is a first pass and will be refined a bit
more as I apply it to the existing User and Finisher searches. I can
add Project search as a drop-in after that.

Also included here are a few annotation and schema changes from main.