Project search and search generalization (#137)
* First pass of search as a Concern

Create a new ActiveModel::Concern top handle the parts of search that
apply to multiple models. This is a first pass and will be refined a bit
more as I apply it to the existing User and Finisher searches. I can
add Project search as a drop-in after that.

Also included here are a few annotation and schema changes from main.

* Relocate project delete button

Move the project delete button to the bottom of the page so it does not
align with the edit button on the previous page.

* Integrate LooseEndsSearchable in User

* Integrate LooseEndsSearchable in Finisher

* Integrate LooseEndsSearchable in Project (Part 1)

Add LooseEndsSearchable concern to Project model and convert the
controller to use the new search method. This is not fully working yet
but it does support most of the features. Still to come are the
Project attributes and then the new fields.

* Wire up project bool attrs in search

* Remove stubbed test class in favor of real tests

* Correct tests & cleanup default sort issues

* Add sub-status search support on Project

* Reflow project search form

* Fix workplace match param

* Support multi-term search as AND

* Add missing sub-status counts

* Add a result count to user search