Assignment last contact date storage
Add the ability to store the last contact date for an assignment. This
will be used to track when the last time a finisher was contacted about
a project. Also included here is a checkbox to update the date when
adding a project note.

This data will be available via the auto CC'd emails in the future. At
that time we can update this field when the email is arrived. This will
let us do that work once instead of looking at email timestamps on the
project search page. This checkbox will give the manager a way to update
if there are out of band contacts as well.