Hotfix: PDF pattern files + login redirect (#340) (#341)
Fix an issue with PDF pattern files. This change uses file.variable? to
guard attempting to generate a variant URL. If the file cannot generate
a variant (like a PDF) we show the file name instead.

Also fixed here was an issue where logins with a return path set in the
Devise cookie were throwing a 500. There was a logic flaw in "&& return"
causing a nil return value. A unit test was added to confirm. The
before_filter used to set return paths needed to be prepended to return
before authentication in the test.

Co-authored-by: Matt Sanford <matt@mzsanford.com>